Replace httpc/sslc transport with curl + mbedTLS
The 3DS's native SSL sysmodule wraps RSA BSAFE SSL-C and tops out at
TLS 1.1 -- libctru exposes exactly three options (default, disable-verify,
force-TLSv1.0), with no TLS 1.2 setting. Its root store holds 11 certs and
contains no ISRG anchor. rommlet's answer was SSLCOPT_DisableVerify on
every request.

Replaces it with devkitPro's 3ds-curl 8.4.0 + 3ds-mbedtls 2.28.8, which
give TLS 1.2 with ECDHE, AES-GCM, SNI and ECDSA -- and let verification
stay ON (CURLOPT_SSL_VERIFYPEER=1, VERIFYHOST=2). The CA bundle comes from
sdmc:/config/ssl/cacert.pem, which Luma3DS already ships and which is
curl's compiled-in default path on this build.

Verification is the actual reason for the swap. httpc is not as limited as
first assumed: httpcAddPostDataRaw/Binary/Ascii give it POST, arbitrary
request headers give it Range, and 3hs (the hShop client) uses exactly
that to implement resumable downloads. What httpc cannot do is verify a
modern certificate chain without the server being tuned for a 2011 TLS
stack -- 3hs ships its own DER cert chain via httpcSelectRootCertChain and
*still* sets SSLCOPT_DisableVerify. curl also spares us hand-rolling
multipart bodies and redirect following.

- http.c/h (new): the whole transport. Brings up the soc service (curl
  needs BSD sockets; rommlet only ever called httpcInit) with a 1MB
  0x1000-aligned buffer. Provides GET, POST-JSON, streaming
  download-to-file, and multipart file upload.
- api.c: static http_get() becomes fetch_json() over the new transport;
  api_download_rom() drops ~150 lines of manual redirect handling and
  chunked httpcDownloadData looping. Zero httpc references remain.
- api_init() now returns bool, since bringing up the network can fail.
- Widen the download and extract progress callbacks from uint32_t to
  uint64_t. Content-Length near the FAT32 4GB ceiling was close enough to
  a uint32 overflow to be worth removing.
- Drop httpcInit/httpcExit from main.c.

Known issue, documented at the call site: curl strips CURLOPT_USERPWD on
cross-host redirects but not custom headers, so the bearer token would
follow a redirect off-host. Harmless against a self-hosted RomM serving
its own files; must be fixed before public exposure.

Verified on hardware against RomM 5.1.0: the request reaches the server
and returns 403 Forbidden, which is what RomM answers for an
unauthenticated /api/platforms -- reproduced identically with curl from a
desktop. Transport confirmed working; authentication is the next phase.

+++ source/zip.h
@@ -10,7 +10,7 @@
 
 // Progress callback for extraction. Returns true to continue, false to cancel.
 // extracted: bytes extracted so far, total: total uncompressed size
-typedef bool (*ExtractProgressCb)(uint32_t extracted, uint32_t total);
+typedef bool (*ExtractProgressCb)(uint64_t extracted, uint64_t total);
 
 // Extract all files from a zip archive into destDir.
 // Deletes the zip file on success.
